* '4.7' of scm.dev.nokia.troll.no:qt/oslo-staging-1: (68 commits)
  Skip test for cocoa: setToolTip
  Replaced qWarning with QCOMPARE/QVERIFY statements in tst_QFileInfo
  Assistant: Handle forward/backward mouse buttons on press on Linux.
  Don't test DBus BIC if QT_NO_DBUS is defined.
  qdoc: Fixed some TOC links.
  Removed broken tests from tst_QFileInfo::isSymlink_data()
  qdoc: Added new table of contents for non-class ref pages
  Revert Merge Request 551. This introduces regressions to Qt.
  Fixed scroll area size calculation on Mac.
  Fix a race where QThread::exit() is "lost" when called after start()
  make a partial build of linguist in no-gui config
  make the code less of a trap
  fix build from top level
  Partially revert MR 543 changes to Linguist.
  Apply f176759fc41abc4cb901c2cbaa15264f2a9ac85b to stdout too.
  Autotest: add some debugging, just in case there's something wrong
  Autotest: fix the fix for the rounding error.
  Fix compile error with QT_NO_LIBRARY in QtMultimedia
  the _setmode() prototype is different on win ce
  qdoc: Changed qdoc to output the new doc format.
  ...

+++ src/gui/widgets/qplaintextedit.cpp
@@ -944,8 +944,8 @@ void QPlainTextEditPrivate::_q_adjustScrollbars()
     int vSliderLength = 0;
     if (!centerOnScroll && q->isVisible()) {
         QTextBlock block = doc->lastBlock();
-        const int visible = static_cast<int>(viewport->rect().height() - margin - 1);
-        int y = 0;
+        const qreal visible = viewport->rect().height() - margin - 1;
+        qreal y = 0;
         int visibleFromBottom = 0;
 
         while (block.isValid()) {
@@ -953,7 +953,7 @@ void QPlainTextEditPrivate::_q_adjustScrollbars()
                 block = block.previous();
                 continue;
             }
-            y += int(documentLayout->blockBoundingRect(block).height());
+            y += documentLayout->blockBoundingRect(block).height();
 
             QTextLayout *layout = block.layout();
             int layoutLineCount = layout->lineCount();
@@ -962,7 +962,7 @@ void QPlainTextEditPrivate::_q_adjustScrollbars()
                 while (lineNumber < layoutLineCount) {
                     QTextLine line = layout->lineAt(lineNumber);
                     const QRectF lr = line.naturalTextRect();
-                    if (int(lr.top()) >= y - visible)
+                    if (lr.top() >= y - visible)
                         break;
                     ++lineNumber;
                 }
